+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Gunpowder Market 1.0.0
 
+- Update to Gunpowder Core 0.3.2 (1.16.4)
+- Notify a seller using the sidebar
+
+---
+# Gunpowder Market 1.0.0
+
 - Ported to standalone module
 - Update to 1.16.2
 - Add `/market expired` to collect expired entries